+++ bin/streamr-stream.ts
@@ -1,0 +1,15 @@
+#!/usr/bin/env node -r ts-node/register
+import { program } from 'commander'
+import pkg from '../package.json'
+
+program
+    .version(pkg.version)
+    .usage('<command> [<args>]')
+    .description('stream subcommands')
+    .command('subscribe', 'subscribe to a stream')
+    .command('publish', 'publish to a stream')
+    .command('list', 'fetch a list of streams')
+    .command('show', 'info about a stream')
+    .command("create", "create a new stream")
+    .command('resend', 'request resend of a stream')
+    .parse(process.argv)