+++ bin/streamr.ts
@@ -4,13 +4,8 @@ import pkg from '../package.json'
 
 program
     .version(pkg.version)
-    .usage('<command> [<args>]')
+    .usage('<command> <subcommand> [<args>]')
     .description('command line tools for interacting with Streamr https://streamr.network')
-    .command('generate', 'generate JSON data')
-    .command('subscribe', 'subscribe to a stream')
-    .command('publish', 'publish to a stream')
-    .command('list', 'fetch a list of streams')
-    .command('show', 'info about a stream')
-    .command("create", "create a new stream")
-    .command('resend', 'request resend of a stream')
+    .command('stream', 'stream subcommands')
+    .command('mock-data', 'mock-data subcommands')
     .parse(process.argv)